chore: rename, remove unused
