+++ backend/user_management/tests.py
@@ -12,77 +12,6 @@ from user_management.models import CustomUser
 User = CustomUser
 
 
-# class FriendsApiTest(TestCase):
-# 	def setUp(self):
-# 		# Create a test user
-# 		self.user = User.objects.create_user(username='testuser', password='testpassword')
-
-# 		# Authenticate the user
-# 		self.client.login(username='testuser', password='testpassword')
-
-# 		self.user2 = User.objects.create_user(username='testuser2', password='testpassword2')
-
-# 	def test_friendships(self):
-# 		response = self.client.get('/users/api/friends/active/', content_type='application/json')
-# 		# self.assertEqual(response.status_code, 301)
-# 		expected_json = {
-# 			'success': True,
-# 			'friends_users': [{}],
-# 		}
-# 		self.assertJSONEqual(json.loads(response.content), expected_json)
-# 		# check existing user
-# 		response = self.client.get(
-# 			'/users/api/friends/request/testuser2/', content_type='application/json'
-# 		)
-# 		self.assertEqual(response.status_code, 200)
-# 		expected_json = {'success': True, 'message': 'Friend request sent successfully.'}
-# 		self.assertJSONEqual(response.content.decode('utf-8'), expected_json)
-# 		# check non-existent user
-# 		response = self.client.get(
-# 			'/users/api/friends/request/testuser3/', content_type='application/json'
-# 		)
-# 		self.assertJSONEqual(
-# 			response.content.decode('utf-8'),
-# 			{'success': False, 'message': 'The target user does not exist!'},
-# 		)
-# 		self.client.logout()
-# 		self.client.login(username='testuser2', password='testpassword2')
-# 		response = self.client.get('/users/api/friends/inactive/', content_type='application/json')
-# 		self.assertJSONEqual(
-# 			response.content.decode('utf-8'),
-# 			{'success': True, 'received': ['testuser'], 'sent': []},
-# 		)
-# 		response = self.client.get(
-# 			'/users/api/friends/accept/testuser/', content_type='application/json'
-# 		)
-# 		self.assertJSONEqual(
-# 			response.content.decode('utf-8'),
-# 			{'success': True, 'message': 'Friend request accepted.'},
-# 		)
-# 		response = self.client.get('/users/api/friends/active/', content_type='application/json')
-# 		self.assertJSONEqual(
-# 			response.content.decode('utf-8'), {'success': True, 'friends_users': ['testuser']}
-# 		)
-# 		response = self.client.get(
-# 			'/users/api/friends/active/testuser/', content_type='application/json'
-# 		)
-# 		self.assertJSONEqual(
-# 			response.content.decode('utf-8'), {'success': True, 'friends_users': ['testuser2']}
-# 		)
-# 		self.client.logout()
-# 		self.client.login(username='testuser', password='testpassword')
-# 		response = self.client.get(
-# 			'/users/api/friends/remove/testuser2/', content_type='application/json'
-# 		)
-# 		self.assertJSONEqual(
-# 			response.content.decode('utf-8'), {'success': True, 'message': 'Friendship ended.'}
-# 		)
-# 		response = self.client.get('/users/api/friends/active/', content_type='application/json')
-# 		self.assertJSONEqual(
-# 			response.content.decode('utf-8'), {'success': True, 'friends_users': []}
-# 		)
-
-
 class BlockManagerTest(TestCase):
 	def setUp(self):
 		# Create test users
